test(gapic): consolidate async otel tests into test_method_async and share fixtures
