+++ packages/google-api-core/tests/conftest.py
@@ -13,6 +13,8 @@
 # limitations under the License.
 
 import os
+import sys
+import types
 from unittest import mock
 
 import pytest
@@ -29,3 +31,36 @@ def mock_mtls_env():
         },
     ):
         yield
+
+
+@pytest.fixture
+def mock_otel(monkeypatch):
+    """Provides a mocked OpenTelemetry environment with tracing enabled."""
+    monkeypatch.setenv("GOOGLE_SDK_EXPERIMENTAL_PYTHON_TRACING_ENABLED", "true")
+    mock_span = mock.MagicMock()
+    mock_tracer = mock.MagicMock()
+    mock_tracer.start_as_current_span.return_value.__enter__.return_value = mock_span
+
+    mock_trace = mock.Mock()
+    mock_trace.get_tracer.return_value = mock_tracer
+    mock_trace.SpanKind.CLIENT = "CLIENT"
+    mock_trace.StatusCode.ERROR = "ERROR"
+
+    with (
+        mock.patch(
+            "google.api_core._observability.is_otel_capabilities_enabled",
+            return_value=True,
+        ),
+        mock.patch.dict(
+            sys.modules,
+            {
+                "opentelemetry": mock.Mock(trace=mock_trace),
+                "opentelemetry.trace": mock_trace,
+            },
+        ),
+    ):
+        yield types.SimpleNamespace(
+            trace=mock_trace,
+            tracer=mock_tracer,
+            span=mock_span,
+        )
